server/requirements.txt
pipecat-ai[google,silero,smallwebrtc]==0.0.103
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.4.1
langsmith>=0.3.0
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
loguru==0.7.3
slack-bolt==1.22.0
google-generativeai==0.8.5
pytest==8.3.4
pytest-asyncio==0.24.0
